# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CGS Games API",
    "version": "1.0.0",
    "description": "Discover community Card Game Simulator games. Public reads require no authentication. Publishing and deletion use Firebase ID tokens and a user profile with a username. API errors retain the legacy error string and add code, message, and hint. GET routes also support HEAD; OPTIONS lists allowed methods. Unknown API paths return JSON 404; unsupported methods return JSON 405 with Allow."
  },
  "servers": [
    {
      "url": "https://cgs.games"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/api/gatcg_spoilers": {
      "get": {
        "operationId": "getGrandArchiveSpoilers",
        "summary": "Read cached Grand Archive spoiler card data",
        "description": "Normally reads the assembled cache. Rate limited to 10 requests per minute per client IP per running instance. Operators may use warm=1 with the separate GATCG_WARM_TOKEN Bearer secret in production; a Firebase ID token does not grant this access. Unauthorized nocache=1 is ignored. warm runs can take up to 240 seconds and are not rate limited.",
        "parameters": [
          {
            "name": "warm",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            },
            "description": "Operator-only cache rebuild; requires the warmToken security scheme in production."
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            },
            "description": "Re-run OCR from scratch. Only applies with warm=1."
          },
          {
            "name": "nocache",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            },
            "description": "Operator-only assembled cache bypass, preserving the OCR cache. Ignored without operator authorization."
          }
        ],
        "security": [
          {},
          {
            "warmToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Spoiler cards in a data array. Pending OCR may leave effect_raw empty.",
            "headers": {
              "X-Cache": {
                "schema": {
                  "type": "string",
                  "enum": ["hit", "miss", "warm", "stale"]
                },
                "description": "Cache result."
              },
              "X-Ocr-Pending": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                },
                "description": "Cards awaiting OCR."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SpoilerCard"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/proxy/{url}": {
      "parameters": [
        {
          "name": "url",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Upstream host and path without a scheme, for example example.com/assets/game.json. This catch-all path can contain slashes; always uses HTTPS. GET forwards query parameters."
        }
      ],
      "get": {
        "operationId": "proxyGet",
        "summary": "Fetch an upstream HTTPS resource",
        "responses": {
          "200": {
            "description": "Upstream body and Content-Type; may be JSON, an image, or other content.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "proxyPost",
        "summary": "Forward a JSON body to an upstream HTTPS resource",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful upstream response, including its headers and body.",
            "content": {
              "*/*": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase Auth ID token from the signed-in user. Requires a creator username."
      },
      "warmToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator GATCG_WARM_TOKEN secret. Only for spoiler cache maintenance; never a Firebase ID token."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "code", "message", "hint"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Legacy error message retained for existing clients."
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
...
      "SpoilerCard": {
        "type": "object",
        "required": [
          "uuid",
          "name",
          "card_image_url",
          "back_card_name",
          "back_card_image_url",
          "types",
          "element",
          "effect_raw"
        ],
        "properties": {
          "uuid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "card_image_url": {
            "type": "string"
          },
          "back_card_name": {
            "type": "string"
          },
          "back_card_image_url": {
            "type": "string"
          },
          "element": {
            "type": "string"
          },
          "effect_raw": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}